Update lab-entering-commands.txt
Hello, Professor, I'm leaving this message here for you to let you know that I was having some trouble with certain questions and was wondering if you could help. I would go to the website we used last week, but I'm having trouble finding my copy of the syllabus. So I tried the best I could to answer the questions given. See you on Monday.